+++ src/main/java/kr/co/morandi/backend/member_management/application/config/oauth/JwtProvider.java
@@ -6,10 +6,10 @@ import jakarta.servlet.http.HttpServletRequest;
 import kr.co.morandi.backend.common.exception.MorandiException;
 import kr.co.morandi.backend.common.exception.errorcode.OAuthErrorCode;
 import kr.co.morandi.backend.member_management.domain.model.member.Member;
-import kr.co.morandi.backend.member_management.domain.model.oauth.constants.Role;
-import kr.co.morandi.backend.member_management.domain.model.oauth.constants.TokenType;
-import kr.co.morandi.backend.member_management.domain.model.oauth.security.SecurityConstants;
-import kr.co.morandi.backend.member_management.domain.model.oauth.response.AuthenticationToken;
+import kr.co.morandi.backend.member_management.infrastructure.oauth.constants.Role;
+import kr.co.morandi.backend.member_management.infrastructure.oauth.constants.TokenType;
+import kr.co.morandi.backend.member_management.infrastructure.security.SecurityConstants;
+import kr.co.morandi.backend.member_management.infrastructure.oauth.response.AuthenticationToken;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
