+++ src/main/java/kr/co/morandi/backend/member_management/application/port/in/oauth/AuthenticationUseCase.java
@@ -1,7 +1,6 @@
 package kr.co.morandi.backend.member_management.application.port.in.oauth;
 
-import jakarta.servlet.http.Cookie;
-import kr.co.morandi.backend.member_management.domain.model.oauth.response.AuthenticationToken;
+import kr.co.morandi.backend.member_management.infrastructure.oauth.response.AuthenticationToken;
 
 public interface AuthenticationUseCase {
     AuthenticationToken getAuthenticationToken(String type, String authenticationCode);
