+++ src/main/java/kr/co/morandi/backend/member_management/application/service/oauth/OAuthService.java
@@ -1,5 +1,5 @@
-package kr.co.morandi.backend.member_management.domain.service.oauth;
-import kr.co.morandi.backend.member_management.domain.model.oauth.OAuthUserInfo;
+package kr.co.morandi.backend.member_management.application.service.oauth;
+import kr.co.morandi.backend.member_management.infrastructure.oauth.OAuthUserInfo;
 
 public interface OAuthService {
     String getType();
