--- src/main/java/kr/co/morandi/backend/member_management/domain/model/oauth/constants/Role.java
@@ -1,5 +1,0 @@
-package kr.co.morandi.backend.member_management.domain.model.oauth.constants;
-
-public enum Role {
-    USER, ADMIN
-}
