+++ src/main/java/kr/co/morandi/backend/member_management/domain/service/oauth/OAuthServiceFactory.java
@@ -1,6 +1,6 @@
 package kr.co.morandi.backend.member_management.domain.service.oauth;
 
-import kr.co.morandi.backend.member_management.domain.service.oauth.OAuthService;
+import kr.co.morandi.backend.member_management.application.service.oauth.OAuthService;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
