+++ src/main/java/kr/co/morandi/backend/member_management/infrastructure/adapter/member/MemberAdapter.java
@@ -1,11 +1,10 @@
 package kr.co.morandi.backend.member_management.infrastructure.adapter.member;
 
-import io.lettuce.core.ScriptOutputType;
 import kr.co.morandi.backend.common.exception.MorandiException;
 import kr.co.morandi.backend.common.exception.errorcode.OAuthErrorCode;
 import kr.co.morandi.backend.member_management.application.port.out.member.MemberPort;
 import kr.co.morandi.backend.member_management.domain.model.member.Member;
-import kr.co.morandi.backend.member_management.domain.model.oauth.constants.SocialType;
+import kr.co.morandi.backend.member_management.infrastructure.oauth.constants.SocialType;
 import kr.co.morandi.backend.member_management.infrastructure.persistence.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
