+++ src/main/java/kr/co/morandi/backend/member_management/infrastructure/filter/oauth/JwtExceptionFilter.java
@@ -9,7 +9,6 @@ import kr.co.morandi.backend.common.exception.MorandiException;
 import kr.co.morandi.backend.common.exception.errorcode.OAuthErrorCode;
 import kr.co.morandi.backend.common.exception.errorcode.global.ErrorCode;
 import kr.co.morandi.backend.common.exception.response.ErrorResponse;
-import kr.co.morandi.backend.member_management.domain.model.oauth.constants.TokenType;
 import kr.co.morandi.backend.member_management.infrastructure.config.CookieUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +20,7 @@ import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
 
-import static kr.co.morandi.backend.member_management.domain.model.oauth.constants.TokenType.REFRESH_TOKEN;
+import static kr.co.morandi.backend.member_management.infrastructure.oauth.constants.TokenType.REFRESH_TOKEN;
 
 @Component
 @RequiredArgsConstructor
