+++ src/main/java/kr/co/morandi/backend/member_management/infrastructure/oauth/constants/Role.java
@@ -1,0 +1,5 @@
+package kr.co.morandi.backend.member_management.infrastructure.oauth.constants;
+
+public enum Role {
+    USER, ADMIN
+}
