+++ src/main/java/kr/co/morandi/backend/member_management/infrastructure/oauth/response/AuthenticationToken.java
@@ -1,4 +1,4 @@
-package kr.co.morandi.backend.member_management.domain.model.oauth.response;
+package kr.co.morandi.backend.member_management.infrastructure.oauth.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
