+++ src/main/java/kr/co/morandi/backend/member_management/infrastructure/security/OAuthDetails.java
@@ -1,4 +1,4 @@
-package kr.co.morandi.backend.member_management.domain.model.oauth.security;
+package kr.co.morandi.backend.member_management.infrastructure.security;
 
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
