+++ src/main/java/kr/co/morandi/backend/member_management/infrastructure/security/SecurityConstants.java
@@ -1,4 +1,4 @@
-package kr.co.morandi.backend.member_management.domain.model.oauth.security;
+package kr.co.morandi.backend.member_management.infrastructure.security;
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
