+++ src/main/java/kr/co/morandi/backend/member_management/infrastructure/security/SecurityUtils.java
@@ -1,4 +1,4 @@
-package kr.co.morandi.backend.member_management.domain.model.oauth.security;
+package kr.co.morandi.backend.member_management.infrastructure.security;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
